Refactor: improve multipart request handling
